archive/Q-Shield/iot-fleet/requirements.txt
paho-mqtt>=1.6.1
liboqs-python>=0.10.0
requests>=2.31.0
